Major bug fix: unconsistent feature space
- When multiple features types were used, iterating over the set of
features was done in a random order, since they are defined as sets
instead of lists. This would cause the feature space to be differenty
order from training to testing, eventhough the specified feature set was
the same! This was fixed by iterating over the sorted set instead of
directly the set
- Removed the nopesq, nomse and save options in evaluation script
- Using the new seed value option from config.py in dataset creation
script